+++ lib/easy_auth/controllers/password_reset.rb
@@ -10,7 +10,7 @@ module EasyAuth::Controllers::PasswordReset
   end
 
   def create
-    if @identity = EasyAuth.find_identity_model(params).where.contains(:uid => [params[:identities_password][:uid]]).first
+    if @identity = EasyAuth.find_identity_model(params).where(:uid => params[:identities_password][:uid]).first
       unencrypted_reset_token = @identity.generate_reset_token!
       PasswordResetMailer.reset(@identity.id, unencrypted_reset_token).deliver
     else
@@ -22,7 +22,7 @@ module EasyAuth::Controllers::PasswordReset
   end
 
   def update
-    @identity = @account.password_identity
+    @identity = @account.password_identities.first
 
     if @account.update_attributes(account_params)
       after_successful_password_reset