+++ spec/dummy/db/schema.rb
@@ -13,11 +13,8 @@
 
 ActiveRecord::Schema.define(version: 20120227014024) do
 
-  # These are extensions that must be enabled in order to support this database
-  enable_extension "plpgsql"
-
   create_table "identities", force: true do |t|
-    t.string   "uid",                   default: [], array: true
+    t.string   "uid"
     t.string   "token"
     t.string   "account_type"
     t.integer  "account_id"
@@ -28,7 +25,7 @@ ActiveRecord::Schema.define(version: 20120227014024) do
     t.string   "reset_token_digest"
   end
 
-  add_index "identities", ["uid"], name: "index_identities_on_uid", using: :btree
+  add_index "identities", ["uid"], name: "index_identities_on_uid"
 
   create_table "users", force: true do |t|
     t.string   "email"